lisp NIL-is-list handling
In IMAP, NIL is not the same as (), but in Common Lisp it is.  When
writing NIL to the wire, handle the CONS-OR-ATOM condition at the
response level to specify how it should be written.  Not handling this
condition is an error.

Provided are helper functions AS-CONS and AS-ATOM to invoke the
appropriate restart.