refactor part 2: use message processor library
Attach the message loop routines to the client connection class.

TODO::

MAKE-RESPONSE is going to use a new kind of IDO lookup that hasn't been
committed yet.  That will follow this refactor.

M-P:SEND-DATA method is just a stub.  You can see how it will invoke the
writer side of the protocol (currently in ``core/connection'').