connections use message processor + dispatch function
Sending data to an IMAP connection using the message processor library
causes the function CORE:WRITE-IMAP4-OBJECT to be invoked with the data.
If the datum is a list, the CAR is the object to be written and the CDR
is the options.

Currently, numbers and non-literal strings are implemented.

(Unimplemented as of this commit: To send delimited lists, the list
needs to be the car of a list.)