default handlers for CAPABILITY and status responses
The standard says we can ignore an unexpected CAPABILITY response, so we
do just that.

As for status responses, those will always be handled directly by
whatever command caused them to be signalled.  We're defining a method
here for when the status response is one of the tagged ones that's
always sent in response to a command so NO-APPLICABLE-METHOD isn't
called.